🚧 Parse markdown dynamically.
Before, the full text was run with each regular expression and any matches were converted to references.

Now, when a match is found, the correlating text is replaced with `'~' × length` . That way it can not match anything further.
Then, any text in the match that can contain more markdown is searched with a new MarkdownParser.